consolidate project queries; resolve public image URLs; nav copy; seed stores image keys
